requirements.txt
# TKLocalAI - Local AI Assistant with LoRA Fine-tuning and RAG
# Optimized for RTX 3050 (8GB VRAM) + 16GB RAM

# ============================================
# Core ML/AI Dependencies
# ============================================

# LLM Inference (GGUF format, GPU accelerated)
llama-cpp-python>=0.2.90

# Transformers ecosystem (for fine-tuning)
torch>=2.1.0
transformers>=4.44.0
accelerate>=0.27.0
datasets>=2.18.0

# Scientific computing (required by sentence-transformers)
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ============================================
# RAG & Embeddings
# ============================================

# Embedding models
sentence-transformers>=2.5.0
huggingface-hub>=0.20.0
tokenizers>=0.15.0
safetensors>=0.4.0

# Vector database
qdrant-client>=1.9.0

# Document processing
pypdf>=4.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
unstructured>=0.12.0

# Text processing
tiktoken>=0.6.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# ============================================
# API Server
# ============================================

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
sse-starlette>=2.0.0

# ============================================
# Desktop UI
# ============================================

pywebview>=5.0.0

# ============================================
# Utilities
# ============================================

python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
tqdm>=4.66.0
httpx>=0.27.0
aiofiles>=23.2.0
watchdog>=4.0.0
loguru>=0.7.0

# ============================================
# Development (optional)
# ============================================

pytest>=8.0.0
black>=24.0.0
ruff>=0.3.0
